Refactoring examples start char A
Refactoring examples start char A.
1. Add ContentPage and refactoring xaml.
2. Remove Unnecessary code.
3. Renamed **Test.cs, **TestPage.xaml, **TestPage.xaml.cs
4. Move Xaml from res/layout to each Test.
5. Unified similar group test.
6. Add comments.
