+++ NUITizenGallery/Examples/AbsoluteLayoutTest/AbsoluteLayoutTest.cs
@@ -21,24 +21,20 @@ using Tizen.NUI.Components;
 
 namespace NUITizenGallery
 {
-    internal class ActivityIndicator : IExample
+    internal class AbsoluteLayoutTest : IExample
     {
         Window window;
-        ActivityIndicatorPage page;
 
         public void Activate()
         {
             Console.WriteLine($"@@@ this.GetType().Name={this.GetType().Name}, Activate()");
             window = NUIApplication.GetDefaultWindow();
-
-            page = new ActivityIndicatorPage();
-            window.Add(page);
+            window.GetDefaultNavigator().Push(new AbsoluteLayoutTestPage());
         }
         public void Deactivate()
         {
             Console.WriteLine($"@@@ this.GetType().Name={this.GetType().Name}, Deactivate()");
-            page.Unparent();
-            page.Dispose();
+            window.GetDefaultNavigator().Pop();
         }
     }
 }
