+++ NUITizenGallery/Examples/AbsoluteLayoutTest/AbsoluteLayoutTestPage.xaml
@@ -1,0 +1,26 @@
+﻿<?xml version="1.0" encoding="UTF-8" ?>
+<ContentPage x:Class="NUITizenGallery.AbsoluteLayoutTestPage"
+  xmlns="http://tizen.org/Tizen.NUI/2018/XAML"
+  xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+  WidthSpecification="{Static LayoutParamPolicies.MatchParent}"
+  HeightSpecification="{Static LayoutParamPolicies.MatchParent}">
+
+    <!-- AppBar is top-side bar with navigation content, title, and action. If you not set any contents, back button is automatically added. -->
+    <ContentPage.AppBar>
+        <AppBar x:Name="appBar" Title="AbsoluteLayoutTestPage"/>
+    </ContentPage.AppBar>
+
+    <!-- Content is main placdeholder of ContentPage. Add your content into this view. -->
+    <ContentPage.Content>
+        <View x:Name="ContentView"
+          WidthSpecification="{Static LayoutParamPolicies.MatchParent}"
+          HeightSpecification="{Static LayoutParamPolicies.MatchParent}">
+            <TextLabel Text="Stylish Header" PointSize="12" Position2D="73,130"/>
+            <View BackgroundColor="#FF0000" Position2D="32,108" Size2D="400,5"/>
+            <View BackgroundColor="#00FF00" Position2D="32,120" Size2D="400,5"/>
+            <View BackgroundColor="#0000FF" Position2D="42,100" Size2D="5,100"/>
+            <View BackgroundColor="#FF00FF" Position2D="54,100" Size2D="5,100"/>
+        </View>
+    </ContentPage.Content>
+
+</ContentPage>
