+++ NUITizenGallery/Examples/ActivityIndicatorTest/ActivityIndicatorTestPage.xaml.cs
@@ -14,17 +14,22 @@
  * limitations under the License.
  *
  */
-using Tizen.NUI;
 using Tizen.NUI.BaseComponents;
 using Tizen.NUI.Components;
 
 namespace NUITizenGallery
 {
-    public partial class AbsoluteLayoutPage : View
+    public partial class ActivityIndicatorTestPage : ContentPage
     {
-        public AbsoluteLayoutPage()
+        public ActivityIndicatorTestPage()
         {
             InitializeComponent();
+
+            string[] imageArray = new string[13];
+            for(int i = 0; i < 13; ++i) {
+                imageArray[i] = Tizen.Applications.Application.Current.DirectoryInfo.Resource + "/images/progress_" + i.ToString() + ".png";
+            }
+            Spinner.ImageArray = imageArray;
         }
     }
 }
