+++ NUITizenGallery/Examples/AlertDialogTest/AlertDialogTestPage.xaml
@@ -1,21 +1,19 @@
 ﻿<?xml version="1.0" encoding="UTF-8" ?>
-<comp:ContentPage x:Class="NUITizenGallery.AlertDialogTestPage"
+<ContentPage x:Class="NUITizenGallery.AlertDialogTestPage"
   xmlns="http://tizen.org/Tizen.NUI/2018/XAML"
-  xmlns:nui="clr-namespace:Tizen.NUI;assembly=Tizen.NUI"
-  xmlns:base="clr-namespace:Tizen.NUI.BaseComponents;assembly=Tizen.NUI"
-  xmlns:comp="clr-namespace:Tizen.NUI.Components;assembly=Tizen.NUI.Components"
   xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-  WidthSpecification="{Static LayoutParamPolicies.MatchParent}" 
+  WidthSpecification="{Static LayoutParamPolicies.MatchParent}"
   HeightSpecification="{Static LayoutParamPolicies.MatchParent}" >
 
+    <!-- AppBar is top-side bar with navigation content, title, and action. If you not set any contents, back button is automatically added. -->
     <ContentPage.AppBar>
-        <AppBar x:Name="appBar"
-                    Title="AlertDialogTestPage"/>
+        <AppBar x:Name="appBar" Title="AlertDialogTestPage"/>
     </ContentPage.AppBar>
 
+    <!-- Content is main placdeholder of ContentPage. Add your content into this view. -->
     <ContentPage.Content>
-        <View x:Name="content"
-                WidthSpecification="{Static LayoutParamPolicies.MatchParent}" 
+        <View x:Name="ContentView"
+                WidthSpecification="{Static LayoutParamPolicies.MatchParent}"
                 HeightSpecification="{Static LayoutParamPolicies.MatchParent}" >
             <View.Layout>
                 <LinearLayout LinearOrientation="Vertical" CellPadding="0,20"/>
@@ -39,4 +37,4 @@
         </View>
     </ContentPage.Content>
 
-</comp:ContentPage>
+</ContentPage>
