--- NUITizenGallery/res/layout/AbsoluteLayoutPage.xaml
@@ -1,16 +1,0 @@
-﻿<?xml version="1.0" encoding="UTF-8" ?>
-<b:View x:Class="NUITizenGallery.AbsoluteLayoutPage"
-  xmlns="http://tizen.org/Tizen.NUI/2018/XAML"
-  xmlns:b="clr-namespace:Tizen.NUI.BaseComponents;assembly=Tizen.NUI"
-  xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    PositionUsesPivotPoint="True" ParentOrigin="Center" 
-    PivotPoint="Center" WidthResizePolicy="FillToParent" 
-    HeightResizePolicy="FillToParent" BackgroundColor="#FFFFFF">
-
-    <b:TextLabel Text="Stylish Header" PointSize="12" Position2D="73,130"/>
-    <b:View BackgroundColor="#FF0000" Position2D="32,108" Size2D="400,5"/>
-    <b:View BackgroundColor="#00FF00" Position2D="32,120" Size2D="400,5"/>
-    <b:View BackgroundColor="#0000FF" Position2D="42,100" Size2D="5,100"/>
-    <b:View BackgroundColor="#FF00FF" Position2D="54,100" Size2D="5,100"/>
-
-</b:View>
