--- NUITizenGallery/res/layout/ActivityIndicatorPage.xaml
@@ -1,12 +1,0 @@
-﻿<?xml version="1.0" encoding="UTF-8" ?>
-<base:View x:Class="NUITizenGallery.ActivityIndicatorPage"
-  xmlns="http://tizen.org/Tizen.NUI/2018/XAML"
-  xmlns:base="clr-namespace:Tizen.NUI.BaseComponents;assembly=Tizen.NUI"
-  xmlns:comp="clr-namespace:Tizen.NUI.Components;assembly=Tizen.NUI.Components"
-  xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    PositionUsesPivotPoint="True" ParentOrigin="Center" 
-    PivotPoint="Center" WidthResizePolicy="FillToParent" 
-    HeightResizePolicy="FillToParent" BackgroundColor="#FFFFFF">
-
-    <comp:Loading x:Name="Spinner" Size2D="100, 100" Position2D="10, 10"/>
-</base:View>
