--- NUITizenGallery/res/layout/AnimationTest1Page.xaml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<base:View x:Class="NUITizenGallery.AnimationTest1Page"
-  xmlns="http://tizen.org/Tizen.NUI/2018/XAML"
-  xmlns:base="clr-namespace:Tizen.NUI.BaseComponents;assembly=Tizen.NUI"
-  xmlns:comp="clr-namespace:Tizen.NUI.Components;assembly=Tizen.NUI.Components"
-  xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    PositionUsesPivotPoint="True" 
-	ParentOrigin="Center" 
-    PivotPoint="Center" 
-	WidthResizePolicy="FillToParent" 
-    HeightResizePolicy="FillToParent" 
-	BackgroundColor="#FFFFFF">
-
-    <base:View.Layout>
-        <LinearLayout LinearOrientation="Vertical" CellPadding="10,10" LinearAlignment="Top"/>
-    </base:View.Layout>
-
-    <base:View x:Name="imgView" WidthResizePolicy="FillToParent" HeightResizePolicy="FillToParent" Weight="0.3">
-        <base:View.Layout>
-            <LinearLayout LinearOrientation="Vertical" LinearAlignment="CenterHorizontal" CellPadding="10,10"/>
-        </base:View.Layout>
-        <ImageView x:Name="imageView" ResourceUrl="*Resource*/images/a.jpg" />
-    </base:View>
-
-    <base:View x:Name="btnView" WidthResizePolicy="FillToParent" HeightResizePolicy="FillToParent" Weight="0.7">
-        <View.Layout>
-            <LinearLayout LinearOrientation="Vertical" LinearAlignment="CenterHorizontal" CellPadding="10,10" />
-        </View.Layout>
-        <comp:Button x:Name="image1Btn" Text="Start Animation" Size2D="680,100"/>
-        <comp:Button x:Name="image2Btn" Text="Stop Animation"  Size2D="680,100"/>
-        <base:TextLabel x:Name="desc1"  Text="" PointSize="15" TextColor="Black"/>
-    </base:View>
-
-
-</base:View>
-
